Lab 2 solutions

+++ src/main/java/miu/edu/waa_lab1/WaaLab1Application.java
@@ -13,7 +13,7 @@ public class WaaLab1Application {
 	}
 
 	@Bean
-	public long longBean() {
+	public int longBean() {
 		return 1;
 	}
